requirements.txt
alembic==1.4.3
APScheduler==3.6.3
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
geographiclib==1.50
geopy==2.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
